pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wordvoyage"
version = "0.1.0"
description = "Daily Bluesky bot for unique words across languages"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "anthropic>=0.60.0",
  "atproto>=0.0.56",
  "Pillow>=10.0.0",
  "psycopg[binary]>=3.2.0",
  "pydantic>=2.8.0",
  "python-dotenv>=1.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

